backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
supabase==2.4.2
python-dotenv==1.0.1
transformers==4.40.0
torch==2.2.2
torchaudio==2.2.2
deepface==0.0.93
opencv-python==4.8.0.76
numpy==1.26.4
textblob==0.18.0
librosa==0.10.1
websockets==12.0
python-jose==3.3.0
httpx==0.27.0
Pillow==10.3.0
pydantic==2.7.0
soundfile==0.12.1
